requirements.txt
# Core dependencies
streamlit>=1.22.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
altair>=5.0.0

# Dataset sources
kaggle>=1.5.13
huggingface-hub>=0.16.0
datasets>=2.13.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.4
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Testing
pytest>=7.3.1
